+++ packages/create/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @sap-ux/create
 
+## 0.13.24
+
+### Patch Changes
+
+-   @sap-ux/abap-deploy-config-inquirer@1.4.8
+-   @sap-ux/adp-tooling@0.14.12
+-   @sap-ux/app-config-writer@0.6.8
+-   @sap-ux/preview-middleware@0.20.19
+-   @sap-ux/system-access@0.6.2
+-   @sap-ux/flp-config-inquirer@0.3.12
+-   @sap-ux/abap-deploy-config-writer@0.1.4
+
 ## 0.13.23
 
 ### Patch Changes