+++ packages/odata-service-inquirer/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @sap-ux/odata-service-inquirer
 
+## 2.4.11
+
+### Patch Changes
+
+-   Updated dependencies [61d4060]
+    -   @sap-ux/axios-extension@1.21.2
+
 ## 2.4.10
 
 ### Patch Changes