+++ packages/preview-middleware/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @sap-ux/preview-middleware
 
+## 0.20.19
+
+### Patch Changes
+
+-   @sap-ux/adp-tooling@0.14.12
+-   @sap-ux/system-access@0.6.2
+
 ## 0.20.18
 
 ### Patch Changes